Inizio richieste

+++ app/src/main/java/com/example/simello/classiServer/InsertHelpRequestInput.java
@@ -8,13 +8,13 @@ public final class InsertHelpRequestInput extends BaseInput {
     private static final long serialVersionUID = -2794341848919295083L;
 
     private String userId;
-    private long latitude;
-    private long longitude;
+    private double latitude;
+    private double longitude;
     private String message;
 
     public InsertHelpRequestInput(){};
 
-    public InsertHelpRequestInput(String userId, long latitude, long longitude, String message) {
+    public InsertHelpRequestInput(String userId, double latitude, double longitude, String message) {
         this.userId = userId;
         this.latitude = latitude;
         this.longitude = longitude;
@@ -25,11 +25,11 @@ public final class InsertHelpRequestInput extends BaseInput {
         return userId;
     }
 
-    public long getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
-    public long getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
